feat(portfolio): add Numero K to experiences/case studies and refresh home+i18n
