+++ src/app/[locale]/home-ui.tsx
@@ -1,0 +1,165 @@
+'use client';
+
+import type { ReactNode } from 'react';
+import Link from 'next/link';
+import { useLocale } from 'next-intl';
+import { usePathname, useRouter } from 'next/navigation';
+import { motion, useReducedMotion } from 'framer-motion';
+import { Languages } from 'lucide-react';
+
+function cn(...classes: Array<string | false | null | undefined>) {
+  return classes.filter(Boolean).join(' ');
+}
+
+export function LocaleToggle() {
+  const locale = useLocale();
+  const router = useRouter();
+  const pathname = usePathname();
+
+  const nextLocale = locale === 'fr' ? 'en' : 'fr';
+
+  function onToggle() {
+    const hash = typeof window !== 'undefined' ? window.location.hash : '';
+    const nextPathname = pathname.replace(
+      /^\/(en|fr)(?=\/|$)/,
+      `/${nextLocale}`
+    );
+    router.push(`${nextPathname}${hash}`);
+  }
+
+  return (
+    <button
+      type="button"
+      onClick={onToggle}
+      className="inline-flex items-center gap-2 rounded-full border border-neutral-200 bg-white/60 px-3 py-2 text-xs font-medium text-neutral-700 backdrop-blur transition hover:bg-neutral-100/50 dark:border-neutral-800 dark:bg-neutral-900/40 dark:text-neutral-200 dark:hover:bg-neutral-900/60"
+      aria-label={
+        nextLocale === 'fr' ? 'Passer en français' : 'Switch to English'
+      }
+      title={nextLocale === 'fr' ? 'FR' : 'EN'}
+    >
+      <Languages className="h-4 w-4" />
+      <span className="uppercase">{locale}</span>
+      <span className="text-neutral-400 dark:text-neutral-500">→</span>
+      <span className="uppercase">{nextLocale}</span>
+    </button>
+  );
+}
+
+export function Section({
+  id,
+  eyebrow,
+  title,
+  children,
+}: {
+  id: string;
+  eyebrow: string;
+  title: string;
+  children: ReactNode;
+}) {
+  return (
+    <section id={id} className="scroll-mt-24">
+      <div className="mb-7">
+        <p className="text-xs uppercase tracking-[0.2em] text-neutral-500 dark:text-neutral-300">
+          {eyebrow}
+        </p>
+        <h2 className="mt-2 text-2xl font-semibold tracking-tight text-neutral-900 dark:text-neutral-50 sm:text-3xl">
+          {title}
+        </h2>
+      </div>
+      {children}
+    </section>
+  );
+}
+
+export function Reveal({
+  children,
+  className,
+}: {
+  children: ReactNode;
+  className?: string;
+}) {
+  const reduce = useReducedMotion();
+  return (
+    <motion.div
+      className={className}
+      initial={reduce ? { opacity: 1 } : { opacity: 0, y: 12 }}
+      whileInView={reduce ? { opacity: 1 } : { opacity: 1, y: 0 }}
+      viewport={{ once: true, margin: '0px 0px -120px 0px' }}
+      transition={{ duration: 0.5, ease: 'easeOut' }}
+    >
+      {children}
+    </motion.div>
+  );
+}
+
+export function Pill({ children }: { children: ReactNode }) {
+  return (
+    <span className="inline-flex items-center rounded-full border border-neutral-200 bg-white/70 px-3 py-1 text-xs text-neutral-700 backdrop-blur dark:border-neutral-800 dark:bg-neutral-900/40 dark:text-neutral-200">
+      {children}
+    </span>
+  );
+}
+
+export function Card({
+  children,
+  className,
+}: {
+  children: ReactNode;
+  className?: string;
+}) {
+  return (
+    <div
+      className={cn(
+        'rounded-2xl border border-neutral-200 bg-white/70 p-6 shadow-sm backdrop-blur transition dark:border-neutral-800 dark:bg-neutral-900/40',
+        'hover:-translate-y-[1px] hover:shadow-md',
+        className
+      )}
+    >
+      {children}
+    </div>
+  );
+}
+
+export function PrimaryButton({
+  href,
+  icon,
+  children,
+}: {
+  href: string;
+  icon: ReactNode;
+  children: ReactNode;
+}) {
+  return (
+    <Link
+      href={href}
+      className="inline-flex items-center gap-2 rounded-full bg-neutral-900 px-5 py-2.5 text-sm font-medium text-white transition hover:bg-neutral-800 dark:bg-neutral-50 dark:text-neutral-900 dark:hover:bg-neutral-200"
+      target={href.startsWith('http') ? '_blank' : undefined}
+      rel={href.startsWith('http') ? 'noopener noreferrer' : undefined}
+    >
+      {icon}
+      {children}
+    </Link>
+  );
+}
+
+export function GhostButton({
+  href,
+  icon,
+  children,
+}: {
+  href: string;
+  icon: ReactNode;
+  children: ReactNode;
+}) {
+  return (
+    <Link
+      href={href}
+      className="inline-flex items-center gap-2 rounded-full border border-neutral-200 bg-white/60 px-5 py-2.5 text-sm font-medium text-neutral-900 backdrop-blur transition hover:bg-neutral-100/50 dark:border-neutral-800 dark:bg-neutral-900/40 dark:text-neutral-50 dark:hover:bg-neutral-900/60"
+      target={href.startsWith('http') ? '_blank' : undefined}
+      rel={href.startsWith('http') ? 'noopener noreferrer' : undefined}
+    >
+      {icon}
+      {children}
+    </Link>
+  );
+}
